>Your code has been rated at 10.00/10
Changes to pass pylint.

  * Shuffled around the imports
  * disabled `duplicate-code` check - it's useful, but can't be
selectively disabled
  * Fixed variable reuse in `for stack_exclude in stack_excludes_list:`
  * Had to ignore `deploy_ecs_service` - desparately needs refactoring
though
  * Other minor code hygiene changes
